requirements.txt
streamlit_folium==0.4.0
matplotlib==3.5.1
streamlit==1.4.0
numpy==1.19.5
pandas==1.3.5
seaborn==0.11.0
folium==0.12.1.post1
Pillow==9.0.0
